% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{annual_pp}
\alias{annual_pp}
\title{Annual Power Production}
\usage{
annual_pp(df)
}
\value{
a data frame with name of power plant and annual power production
of the power plant.
}
\description{
Function that load data of power plant and power production.
}
