% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_wacc_npv.R
\name{cashflow}
\alias{cashflow}
\title{Cashflow}
\usage{
cashflow(df_revenue, df_costs, tax = 0.22)
}
\arguments{
\item{df_revenue}{df with annual revenue for the company.}

\item{df_costs}{df with annual costparameters for the company.}

\item{tax}{taxrate of the EBIT for the company.}
}
\value{
a tibble including name of company, year and cashflow that
corresponding year
}
\description{
Function that calculates cashflow for the company.
}
