% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue.R
\name{exp_revenue_per_year}
\alias{exp_revenue_per_year}
\title{Calculate revenue}
\usage{
exp_revenue_per_year(production, pricepath)
}
\arguments{
\item{pricepath}{}
}
\value{
revenue per year for the years available in the price path
}
\description{
Function that takes the price path as well as annual production for a power
plant as input, and returns revenue per year for the years available in the
price path.
}
