% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data_annual_pp}
\alias{load_data_annual_pp}
\title{Load data}
\usage{
load_data_annual_pp()
}
\value{
a data from of annual production of power for all power plants
in Norway.
}
\description{
Function that loads data from power plants in Norway.
}
\examples{
df <- load_data_annual_pp()
}
