% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_wacc_npv.R
\name{npv}
\alias{npv}
\title{Net Present Value}
\usage{
npv(cf, wacc)
}
\arguments{
\item{cf}{df with cashflows for the firm.}

\item{wacc}{a number in percent, the real total weighted average cost of
capital (WACC).}
}
\value{
the NPV for a firm at time 0.
}
\description{
Function that calculates the Net Present Value (NPV) of a firm.
}
