% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\docType{data}
\name{pricepath}
\alias{pricepath}
\title{Price of power.}
\format{Data frame with columns
\describe{
\item{date}{date of the registered price.}
\item{power_price}{price of one kw of power.}
}}
\source{
"C:/Users/NO100464/Documents/R/valuationPowerCompanies/data-raw/price.R"
}
\usage{
pricepath
}
\description{
Look up price of power on the given dates.
}
\examples{
  pricepath
}
\keyword{datasets}
