% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_wacc_npv.R
\name{wacc}
\alias{wacc}
\title{WACC

Function that calculates the total weighted average cost of capital (WACC)
for a specific power plant.}
\usage{
wacc(cashflow, asset_beta = 0.6, risk_free_rate = 0.02,
  market_risk_prem = 0.05, other_adj = 0.02, debt_prem = 0.025,
  debt_share = 0.5, corp_income_tax = 0.22, inflation = 0.02)
}
\arguments{
\item{asset_beta}{asseta beta for the company.}

\item{risk_free_rate}{risk-free rate of production for the company.}

\item{market_risk_prem}{Market risk premium for operating for the company.}

\item{other_adj}{Other adjustment for the company.}

\item{debt_prem}{Debt premium for the company.}

\item{debt_share}{Debt share for the company.}

\item{corp_income_tax}{Corporation income tax for the company.}

\item{inflation}{Rate of inflation in the economy.}

\item{cashflows}{estimated future cashflow for the company.}
}
\value{
Returns the real total weighted average cost of capital (WACC) in
percent.
}
\description{
WACC

Function that calculates the total weighted average cost of capital (WACC)
for a specific power plant.
}
\examples{
cashflows <- cashflow(df_revenue = aga_revenue, df_costs = aga_costs)
}
