% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{costs}
\alias{costs}
\title{Calculate costs}
\usage{
costs(pricepath, operating_cost = 0.3, investments = 0.2,
  depreciation = 0.2, annual_powerprod)
}
\arguments{
\item{annual_powerprod}{}
}
\value{
Returns operating costs, investment costs and depreciation per year,
for as many years as are included in the pricepath.
}
\description{
Function that calculates costs per year for a power plant.
}
