% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_wacc_npv.R
\name{wacc}
\alias{wacc}
\title{WACC

Function that calculates the total weighted average cost of capital (WACC)
for a specific power plant.}
\usage{
wacc(cashflows, asset_beta = 0.6, risk_free_rate = 0.02,
  market_risk_prem = 0.05, other_adj = 0.02, debt_prem = 0.025,
  debt_share = 0.5, corp_income_tax = 0.22, inflation = 0.02)
}
\arguments{
\item{inflation}{}
}
\value{
Returns the real total weighted average cost of capital (WACC) in
percent.
}
\description{
WACC

Function that calculates the total weighted average cost of capital (WACC)
for a specific power plant.
}
\examples{
cashflows <- cashflow(df_revenue = aga_revenue, df_costs = aga_costs)
}
